+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 **Features**:
 
+- Added support **MMM** in format for `dateToString, dateFromString`
 - Controls **Number, Date, Calendar** are locale based and depeneds on [`localeInfo`](src/helpers/localeInfo.ts) helper
 - Added elements